style: Files indentation and typing correction

+++ IA/CNN/CNNBase.py
@@ -2,10 +2,12 @@ import keras
 import keras_tuner as kt
 from IA.NNBase import NNBase
 
-class CNNBase( NNBase ):
+
+class CNNBase(NNBase):
     """
     CNN Base Model
     """
+
     def __init__(self, input_shape, window_length: int):
         super().__init__(input_shape, window_length)
         self.convolution_counter: int = 0
@@ -15,23 +17,12 @@ class CNNBase( NNBase ):
         Return the name of the model.
         """
         return "CNN"
-    
-    def create_dropout_layer(self, hyper_param: kt.HyperParameters,
-                             min_value: int, max_value: int,
-                             step_value: int, default_value: int):
+
+    def create_dropout_layer(self, hyper_param: kt.HyperParameters, min_value: float, max_value: float, step_value: float, default_value: float):
         return super().create_dropout_layer(hyper_param, min_value, max_value, step_value, default_value)
 
-    def create_convolution_layer(self, hyper_param: kt.HyperParameters,
-                                 min_value: int, max_value: int,
-                                 step_value: int, default_value: int, padding: str = 'valid'):
-        convolution_layer = keras.layers.Conv2D(
-            hyper_param.Int(name=f"conv_{self.convolution_counter}",
-                            min_value=min_value, max_value=max_value,
-                            step=step_value, default=default_value),
-            kernel_size=(3, 3),
-            activation='relu',
-            padding=padding
-        )
+    def create_convolution_layer(self, hyper_param: kt.HyperParameters, min_value: int, max_value: int, step_value: int, default_value: int, padding: str = "valid"):
+        convolution_layer = keras.layers.Conv2D(hyper_param.Int(name=f"conv_{self.convolution_counter}", min_value=min_value, max_value=max_value, step=step_value, default=default_value), kernel_size=(3, 3), activation="relu", padding=padding)
         self.convolution_counter = self.convolution_counter + 1
         return convolution_layer
 
@@ -42,4 +33,4 @@ class CNNBase( NNBase ):
         self.model.add(super().create_dense_layer(hyper_param, min_value=16, max_value=128, step_value=8, default_value=64))
         self.model.add(super().create_dropout_layer(hyper_param, min_value=0.1, max_value=0.8, step_value=0.1, default_value=0.5))
 
-        self.model.add(keras.layers.Dense(1, activation='sigmoid'))
+        self.model.add(keras.layers.Dense(1, activation="sigmoid"))
